requirements.txt
chainlit
python-dotenv
requests
agents
